Pull asynchronous GCS functions into their own file.
