+++ gubernator/main_test.py
@@ -27,11 +27,14 @@ import unittest
 import urlparse
 
 import webtest
-import webapp2
 
 import cloudstorage as gcs
 
 import main
+import gcs_async
+import gcs_async_test
+
+write = gcs_async_test.write
 
 app = webtest.TestApp(main.app)
 
@@ -64,50 +67,6 @@ def init_build(build_dir, started=True, finished=True):
         write(build_dir + 'finished.json',
               {'result': 'SUCCESS', 'timestamp': 1406536800})
     write(build_dir + 'artifacts/junit_01.xml', JUNIT_SUITE)
-
-
-def add_gcs_json_handler(stub, structure):
-    '''
-    Add a stub to mock out GCS JSON API ListObject requests-- with
-    just enough detail for our code.
-
-    This is based on google.appengine.ext.cloudstorage.stub_dispatcher.
-
-    Args:
-        stub: a URLFetch stub, to register our new handler against.
-        structure: a dictionary of {paths: subdirectory names}.
-            This will be transformed into the (more verbose) form
-            that the ListObject API returns.
-    '''
-    prefixes_for_paths = {}
-
-    for path, subdirs in structure.iteritems():
-        path = 'pr-logs/pull/' + path
-        prefixes_for_paths[path] = ['%s%s/' % (path, d) for d in subdirs]
-
-    def matches(url):
-        return url.startswith(main.STORAGE_API_URL)
-
-    def dispatch(method, url, payload):
-        if method != 'GET':
-            raise ValueError('unhandled method %s' % method)
-        parsed = urlparse.urlparse(url)
-        path = parsed.path
-        param_dict = urlparse.parse_qs(parsed.query, True)
-        prefix = param_dict['prefix'][0]
-        return json.dumps({'prefixes': prefixes_for_paths[prefix]})
-
-    def fetch_stub(url, payload, method, headers, request, response,
-                   follow_redirects=False, deadline=None,
-                   validate_certificate=None):
-        result = dispatch(method, url, payload)
-        response.set_statuscode(200)
-        response.set_content(result)
-        header = response.add_header()
-        header.set_key('content-length')
-        header.set_value(str(len(result)))
-
-    stub._urlmatchers_to_fetch_functions.append((matches, fetch_stub))
 
 
 class HelperTest(unittest.TestCase):
@@ -155,7 +114,7 @@ class AppTest(unittest.TestCase):
         self.testbed.init_datastore_v3_stub()
         init_build(self.BUILD_DIR)
         # redirect GCS calls to the local proxy
-        main.GCS_API_URL = gcs.common.local_api_url()
+        gcs_async.GCS_API_URL = gcs.common.local_api_url()
 
     def get_build_page(self):
         return app.get('/build' + self.BUILD_DIR)
@@ -245,7 +204,7 @@ class AppTest(unittest.TestCase):
         self.assertIn('somejob/">somejob</a>', response)
 
     def init_pr_directory(self):
-        add_gcs_json_handler(self.testbed.get_stub('urlfetch'),
+        gcs_async_test.install_handler(self.testbed.get_stub('urlfetch'),
             {'123/': ['build', 'e2e'],
              '123/build/': ['11', '10', '12'],  # out of order
              '123/e2e/': ['47', '46']})
@@ -276,7 +235,7 @@ class AppTest(unittest.TestCase):
         self.assertIn('github.com/kubernetes/kubernetes/pull/123', response)
 
     def test_pr_handler_missing(self):
-        add_gcs_json_handler(self.testbed.get_stub('urlfetch'),
+        gcs_async_test.install_handler(self.testbed.get_stub('urlfetch'),
             {'124/': []})
         response = app.get('/pr/124')
         self.assertIn('No Results', response)
